Se cambió el ENUM TipoDNI por una clase Entity con su respectiva clase Service, ServiceImpl y Repository. Se completaron los Controller y algunos cambios generales

+++ src/main/java/com/example/demo/servicies/impl/CandidatoServiceImpl.java
@@ -1,6 +1,8 @@
 package com.example.demo.servicies.impl;
 
-import com.example.demo.exceptions.CandidatoNoEncontrado;
+import com.example.demo.exceptions.DniExistenteExcpetion;
+import com.example.demo.exceptions.DniInexistenteExcpetion;
+import com.example.demo.exceptions.IdEncontradoException;
 import com.example.demo.models.enitities.Candidato;
 import com.example.demo.models.views.CandidatoDTO;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -8,6 +10,7 @@ import org.springframework.stereotype.Service;
 import com.example.demo.repositories.CandidatoRepository;
 import com.example.demo.servicies.CandidatoService;
 
+import javax.persistence.EntityExistsException;
 import java.util.List;
 
 @Service
@@ -16,8 +19,11 @@ public class CandidatoServiceImpl implements CandidatoService {
     @Autowired
     CandidatoRepository candidatoRepository;
 
-    // el .save sabe distinguir si es un candidato nuevo. Si lo es, le agrega ID, sino la modifica
     public void guardarCandidato(CandidatoDTO candidatoDTO){
+
+        if(candidatoRepository.findByDni(candidatoDTO.getDni()) != null){
+            throw new DniExistenteExcpetion("Ya existe un candidato con el DNI ingresado.");
+        }
 
         Candidato candidato = Candidato.builder()
                 .nombre(candidatoDTO.getNombre())
@@ -30,29 +36,42 @@ public class CandidatoServiceImpl implements CandidatoService {
          candidatoRepository.save(candidato);
     }
 
+    public void modificarCandidato(CandidatoDTO candidatoDTO){
+        if(candidatoRepository.findByDni(candidatoDTO.getDni()) == null){
+            throw new DniInexistenteExcpetion("No existe un candidato con el DNI ingresado.");
+        }
+
+        Candidato candidato = Candidato.builder()
+                .nombre(candidatoDTO.getNombre())
+                .apellido(candidatoDTO.getApellido())
+                .dni(candidatoDTO.getDni())
+                .fechaNacimiento(candidatoDTO.getFechaNacimiento())
+                .tipoDNI(candidatoDTO.getTipoDNI())
+                .build();
+
+        candidatoRepository.save(candidato);
+    }
+
     public List<Candidato> obtenerCandidatos(){
         return candidatoRepository.findAll();
     }
 
     public void eliminarCandidatoPorId(Long idCandidatoDTO){
+
+        if(candidatoRepository.findById(idCandidatoDTO).orElseThrow() == null){
+            throw new IdEncontradoException("No se encontró un candidato con ese ID.");
+        }else{
             candidatoRepository.deleteById(idCandidatoDTO);
+        }
     }
 
     public Candidato buscarCandidatoPorId(Long idCandidato){
         return candidatoRepository.findById(idCandidato)
-                .orElseThrow(() -> new CandidatoNoEncontrado("No se encontró el candidato")); //Supplier, no recibe parametros porque es un proveedor
+                .orElseThrow(() -> new EntityExistsException("No se encontró el candidato")); //Supplier, no recibe parametros porque es un proveedor
     }
 
-    /*
-    public List<Candidato> findByTecnologia(String tecnologia) {
-        List<Candidato> candidatos = candidatoRepository.findByTecnologia(tecnologia);
-
-        for (Candidato cand: candidatos){
-
-        }
-
-        return null;
+    public Candidato buscarCandidatoPorDni(String dni){
+        return candidatoRepository.findByDni(dni);
     }
-*/
 
 }
