+++ src/main/java/com/example/demo/servicies/impl/TecnologiaServiceImpl.java
@@ -1,6 +1,7 @@
 package com.example.demo.servicies.impl;
 
-import com.example.demo.exceptions.TecnologiaNoEncontrada;
+import com.example.demo.exceptions.IdEncontradoException;
+import com.example.demo.exceptions.IdNoEncontradoException;
 import com.example.demo.models.enitities.Tecnologia;
 import com.example.demo.models.views.TecnologiaDTO;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -8,6 +9,7 @@ import org.springframework.stereotype.Service;
 import com.example.demo.repositories.TecnologiaRepository;
 import com.example.demo.servicies.TecnologiaService;
 
+import javax.persistence.EntityExistsException;
 import java.util.List;
 
 @Service
@@ -16,13 +18,29 @@ public class TecnologiaServiceImpl implements TecnologiaService {
     @Autowired
     TecnologiaRepository tecnologiaRepository;
 
-    // el .save sabe distinguir si es una tecnologia nueva. Si lo es, le agrega ID, sino la modifica
+
     public void guardarTecnologia(TecnologiaDTO tecnologiaDTO){
+        if(tecnologiaDTO.getIdTecnologia() != null){
+            throw new IdEncontradoException("La tecnologia tiene ID. Utilice modificarTecnologia");
+        }
         Tecnologia tecnologia = Tecnologia.builder()
+                .nombre(tecnologiaDTO.getNombre())
                 .version(tecnologiaDTO.getVersion())
                 .build();
 
          tecnologiaRepository.save(tecnologia);
+    }
+
+    public void modificarTecnologia(TecnologiaDTO tecnologiaDTO){
+        if(tecnologiaDTO.getIdTecnologia() == null){
+            throw new IdNoEncontradoException("No se puede modificar una tecnologa sin ID.");
+        }
+        Tecnologia tecnologia = Tecnologia.builder()
+                .nombre(tecnologiaDTO.getNombre())
+                .version(tecnologiaDTO.getVersion())
+                .build();
+
+        tecnologiaRepository.save(tecnologia);
     }
 
     public List<Tecnologia> obtenerTecnologias(){
@@ -31,13 +49,16 @@ public class TecnologiaServiceImpl implements TecnologiaService {
     }
 
     public void eliminarTecnologiaPorId(Long idTecnologia){
-
-        tecnologiaRepository.deleteById(idTecnologia);
+        if(tecnologiaRepository.findById(idTecnologia).orElseThrow() == null){
+            throw new IdEncontradoException("No se encontró una tecnologia con ese ID.");
+        }else{
+            tecnologiaRepository.deleteById(idTecnologia);
+        }
     }
 
     public Tecnologia buscarTecnologiaPorId(Long idTecnologia){
         return tecnologiaRepository.findById(idTecnologia)
-                .orElseThrow(()-> new TecnologiaNoEncontrada("No se encontro la tecnologia"));
+                .orElseThrow(()-> new EntityExistsException("No se encontró el candidato"));
     }
 
 }
